+++ aiohttp_client_cache/backends/base.py
@@ -225,8 +225,12 @@ class CacheBackend:
         async for r in self.responses.values():
             yield r.url  # type: ignore
 
+    async def close(self):
+        """Close any active connections, if applicable"""
+
 
 # TODO: Support yarl.URL like aiohttp does?
+# TODO: Implement __aiter__?
 class BaseCache(metaclass=ABCMeta):
     """A wrapper for lower-level cache storage operations. This is separate from
     :py:class:`.CacheBackend` to allow a single backend to contain multiple cache objects.