+++ test/conftest.py
@@ -1,5 +1,5 @@
 from contextlib import asynccontextmanager
-from datetime import datetime, timezone
+from datetime import datetime
 from logging import basicConfig, getLogger
 from os import getenv
 from tempfile import NamedTemporaryFile
@@ -10,6 +10,7 @@ import pytest
 from aiohttp_client_cache import CachedResponse, CachedSession, SQLiteBackend
 
 ALL_METHODS = ['GET', 'HEAD', 'OPTIONS', 'POST', 'PUT', 'PATCH', 'DELETE']
+CACHE_NAME = 'pytest_cache'
 HTTPBIN_METHODS = ['GET', 'POST', 'PUT', 'PATCH', 'DELETE']
 HTTPBIN_FORMATS = [
     'brotli',