+++ test/integration/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa: F401
+from test.integration.base_backend_test import BaseBackendTest
+from test.integration.base_storage_test import BaseStorageTest