+++ HISTORY.md
@@ -2,6 +2,8 @@
 
 ## 0.5.0 (2021-TBD)
 * Add a filesystem backend
+* Add `RedisBackend.close()` method 
+* Add `MongoDBPickleCache.values()` method that deserializes items
 
 ---
 ### 0.4.3 (2021-07-27)